- Completed basic functions for Cohorts; both repo and controllers.

+++ src/main/java/application/controllers/CohortController.java
@@ -1,11 +1,14 @@
 package application.controllers;
 
+import application.models.Cohorts;
 import application.repositories.CohortRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @RestController
 @RequestMapping("cohort")
@@ -17,8 +20,9 @@ public class CohortController {
     /* CREATORS */
 
     @RequestMapping(value = "/create",  produces = "application/json")
-    public String create(@RequestParam("name") String cohort_name, @RequestParam("desc") String description) {
-        return this.cohortRepository.create(cohort_name, description);
+    public String create(@RequestParam("cname") String cohort_name, @RequestParam(value = "desc", required = false) String description,
+                         @RequestParam(value = "pname", required = false) String parent_name) {
+        return this.cohortRepository.create(cohort_name, description, parent_name);
     }
 
 
@@ -50,7 +54,41 @@ public class CohortController {
     /* REMOVALS */
 
     @RequestMapping(value = {"/remove", "/delete"}, produces = "application/json")
-    public String remove(@RequestParam(value = "cname", required = false) String cohort_name) {
-        return this.cohortRepository.removeCohortByName(cohort_name);
+    public String remove(@RequestParam(value = "cname", required = false) String cohort_name,
+                         @RequestParam(value = "pname", required = false) String parent_name) {
+
+        /* Remove by Cohort Name */
+        if(cohort_name != null) {
+            return this.cohortRepository.removeCohortByName(cohort_name);
+        }
+
+        /* Remove by Parent ID */
+        if(parent_name != null) {
+            return this.cohortRepository.removeCohortByParentID(parent_name);
+        } else { return "You must specify the parameters"; }
+    }
+
+    @RequestMapping(value = "null", produces = "application/json")
+    public String nuller(@RequestParam(value = "cname") String cohort_name,
+                         @RequestParam(value = "target") String target) {
+
+        switch (target) {
+            case "parent": return this.cohortRepository.nullParentID(cohort_name);
+            case "desc": return this.cohortRepository.nullDescription(cohort_name);
+            default: return "Please, use the correct target name: name, parent, or desc";
+        }
+    }
+
+    @RequestMapping(value = "get", produces = "application/json")
+    public List<Cohorts> get(@RequestParam(value = "text") String variable,
+                             @RequestParam(value = "target") String target) {
+
+        switch (target) {
+            case "name": return this.cohortRepository.getCohortByName(variable);
+            case "parent": return this.cohortRepository.getCohortByParentName(variable);
+            case "desc": return this.cohortRepository.getCohortByDescription(variable);
+            case "id": return this.cohortRepository.getCohortByDBID(variable);
+            default: return new ArrayList<>();
+        }
     }
 }
